requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==39.0.0
Deprecated==1.2.13
exceptiongroup==1.1.0
firebase==3.0.1
frozenlist==1.3.3
gcloud==0.18.3
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
iniconfig==2.0.0
jwcrypto==1.4.2
Markdown==3.4.1
markdownify==0.11.6
multidict==6.0.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
protobuf==4.21.12
py-trello==0.18.0
pyairtable==1.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyTelegramBotAPI==4.11.0
pytest==7.2.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2022.7
PyYAML==6.0
redis==4.4.2
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sseclient==0.0.27
telebot-components==0.6.0
telebot-utils==0.0.8
tenacity==8.1.0
tomli==2.0.1
trafaret==2.1.1
trafaret-config==2.0.2
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.13
wrapt==1.14.1
yarl==1.8.2
